Remove v1-specific copy from ProjectDetailsForm
