+++ src/locales/pt/messages.po
@@ -374,7 +374,7 @@ msgstr "Cancelar"
 msgid "Cannot redeem tokens for ETH because this project has no overflow."
 msgstr ""
 
-#: src/components/v1/V1Create/ProjectForm.tsx
+#: src/components/v1/V1Create/index.tsx
 msgid "Changes to these attributes can be made at any time and will be applied to your project immediately."
 msgstr "As mudanças para esses atribuitos podem ser feitas a qualquer momento e irão ser aplicadas ao seu projeto imediatamente."
 
@@ -1139,7 +1139,7 @@ msgid "Project description"
 msgstr "Descrição do projeto"
 
 #: src/components/v1/V1Create/ConfirmDeployProject.tsx
-#: src/components/v1/V1Create/ProjectForm.tsx
+#: src/components/v1/V1Create/index.tsx
 #: src/components/v1/V1Create/index.tsx
 msgid "Project details"
 msgstr "Detalhes do projeto"
@@ -1149,7 +1149,7 @@ msgstr "Detalhes do projeto"
 msgid "Project handle"
 msgstr "Identificador do projeto"
 
-#: src/components/v1/V1Create/ProjectForm.tsx
+#: src/components/shared/forms/ProjectDetailsForm.tsx
 msgid "Project handle must be unique."
 msgstr "O identificador do projeto deve ser único."
 
@@ -1360,7 +1360,7 @@ msgstr "Salvar identificador"
 msgid "Save payouts"
 msgstr "Salvar pagamentos"
 
-#: src/components/v1/V1Create/ProjectForm.tsx
+#: src/components/shared/forms/ProjectDetailsForm.tsx
 msgid "Save project details"
 msgstr "Salvar detalhes do projeto"
